Add script for converting tree-sitter js dsl to ebnf
The transliterated tree-sitter grammar based on the official scala
language spec has some serious issues that make it unlikely to viably
work as a tree-sitter parser. It still has been extremely useful as a
proof of concept that tree-sitter grammar's could be more easily written
using the custom ebnf syntax based on the w3c spec with some tree-sitter
extensions. This commit adds a node.js script that can convert a grammar
written in the tree-sitter dsl into the ebnf format. The output is not
pretty printed because it didn't seem worth the effort since converting
an existing grammar should typically be a one off. I verified that I
could go back and forth between the tree-sitter and ebnf formats with
the grammar defined for scala at
https://github.com/tree-sitter/tree-sitter-scala.

To make this work, I had to add the missing functionality from
tree-sitter into the ebnf dsl. The two formats should be equivalent ways
of writing tree-sitter grammars.