Add fork of tree-sitter-scala
This is effectively a fork of tree-sitter-scala except that I updated it
to use a scala.ebnf file that I generated from the existing grammar.js
using the tree-sitter-to-ebnf.js script instead of directly modifiying
the grammar.js. This should make iterating on the existing grammar and
identifying discrepencies with the official spec much easier.